Add full support for parsing DTs at boot
To recap this absolute horrific milestone, it appears that CFFI can only have
one module at runtime. Since dtlib is coming from zephyr and we want to keep it
separate, I've had to add a "bundling" step that stores the library in a dict,
then gzips/base64s it into a string, bundles that string into the main program
code. On the other side we have a utility class that gets called on startup
that reverses this entire process to unpack everything and does a little extra
magic to insert it into the namespace, ready to be imported by anyone who needs
it.