Add a class registration system to build drivers.
The simulation now parses the DT and classes which inherit from DeviceClass
are allowed to implement a @register_probe function which declares a set of
device_ids which they'll be probed for.

Additionally, a fix was implemented to force loading core and utilities first,
ensuring model code always has them available.