Rewrote the makefile for hygiene
All driver code is now in the system/ dir and outputs are generated in an
output/ dir by default. Both can be overriden with the CODE_DIR and OUT_DIR
environment variables respectively. Additionally, this seems to have fixed an
issue with register_driver not passing in sdrv properly somehow.